Move useLogin to web/graphql/auth module
We should have all queries in one place at the end.
